Moves ObservedCadence class into snat_sim.models.light_curve
