+++ snat_sim/models/light_curve.py
@@ -3,13 +3,16 @@
 from __future__ import annotations
 
 from copy import copy
+from dataclasses import dataclass
 from typing import *
-from typing import TYPE_CHECKING
+from typing import Optional, Union
 
 import numpy as np
 import pandas as pd
 import sncosmo.photdata
 from astropy.table import Table
+
+from .. import types
 
 if TYPE_CHECKING:  # pragma: no cover
     from ..types import Numeric
@@ -18,6 +21,98 @@ SNCOSMO_ALIASES = dict()
 for column_name, alias_set in sncosmo.photdata.PHOTDATA_ALIASES.items():
     for alias in alias_set:
         SNCOSMO_ALIASES[alias] = column_name
+
+
+# noinspection PyAttributeOutsideInit,PyUnresolvedReferences
+@dataclass
+class ObservedCadence:
+    """The observational sampling of an astronomical light-curve
+
+    The zero-point, zero point system, and gain arguments can be a
+    collection of values (one per ``obs_time`` value), or a single value
+    to apply at all observation times.
+
+    Args:
+        obs_times: Array of observation times for the light-curve
+        bands: Array of bands for each observation
+        zp: The zero-point or an array of zero-points for each observation
+        zpsys: The zero-point system or an array of zero-point systems
+        gain: The simulated gain or an array of gain values
+    """
+
+    obs_times: Collection[float]
+    bands: Collection[str]
+    skynoise: types.FloatColl
+    zp: types.FloatColl
+    zpsys: Union[str, Collection[str]]
+    gain: types.FloatColl
+
+    def __eq__(self, other: ObservedCadence) -> bool:
+        attr_list = ['obs_times', 'bands', 'skynoise', 'zp', 'zpsys', 'gain']
+        return np.all(np.equal(getattr(self, attr), getattr(other, attr)) for attr in attr_list)
+
+    @property
+    def skynoise(self) -> np.array:
+        return self._skynoise.copy()
+
+    @skynoise.setter
+    def skynoise(self, skynoise: types.FloatColl):
+        self._skynoise = np.full_like(self.obs_times, skynoise)
+
+    @property
+    def zp(self) -> np.array:
+        return self._zp.copy()
+
+    @zp.setter
+    def zp(self, zp: types.FloatColl):
+        self._zp = np.full_like(self.obs_times, zp)
+
+    @property
+    def zpsys(self) -> np.array:
+        return self._zpsys.copy()
+
+    @zpsys.setter
+    def zpsys(self, zpsys: types.FloatColl):
+        self._zpsys = np.full_like(self.obs_times, zpsys, dtype='U8')
+
+    @property
+    def gain(self) -> np.array:
+        return self._gain.copy()
+
+    @gain.setter
+    def gain(self, gain: types.FloatColl):
+        self._gain = np.full_like(self.obs_times, gain)
+
+    def to_sncosmo(self) -> Table:
+        """Return the observational cadence as an ``astropy.Table``
+
+        The returned table of observations is formatted for use with with
+        the ``sncosmo`` package.
+
+        Returns:
+            An astropy table representing the observational cadence in ``sncosmo`` format
+        """
+
+        observations = Table(
+            {
+                'time': self.obs_times,
+                'band': self.bands,
+                'gain': self.gain,
+                'skynoise': self.skynoise,
+                'zp': self.zp,
+                'zpsys': self.zpsys
+            },
+            dtype=[float, 'U1000', float, float, float, 'U100']
+        )
+
+        observations.sort('time')
+        return observations
+
+    def __repr__(self) -> str:  # pragma: no cover
+        repr_list = self.to_sncosmo().__repr__().split('\n')
+        repr_list[0] = super(ObservedCadence, self).__repr__()
+        repr_list.pop(2)
+        return '\n'.join(repr_list)
 
 
 class LightCurve:
